refactor(backend): Change params type of dto to long and make error message more precise  #1644

+++ backend/src/main/java/ch/puzzle/okr/dto/EvaluationDto.java
@@ -1,7 +1,7 @@
 package ch.puzzle.okr.dto;
 
-public record EvaluationDto(int objectiveAmount, int completedObjectivesAmount,
-        int successfullyCompletedObjectivesAmount, int keyResultAmount, int keyResultsOrdinalAmount,
-        int keyResultsMetricAmount, int keyResultsInTargetOrStretchAmount, int keyResultsInFailAmount,
-        int keyResultsInCommitAmount, int keyResultsInTargetAmount, int keyResultsInStretchAmount) {
+public record EvaluationDto(long objectiveAmount, long completedObjectivesAmount,
+        long successfullyCompletedObjectivesAmount, long keyResultAmount, long keyResultsOrdinalAmount,
+        long keyResultsMetricAmount, long keyResultsInTargetOrStretchAmount, long keyResultsInFailAmount,
+        long keyResultsInCommitAmount, long keyResultsInTargetAmount, long keyResultsInStretchAmount) {
 }
